Fixed the database struct like the diagram
